+++ js/src/index.js
@@ -4,3 +4,4 @@ export { Parser } from './Parser.js';
 export { ParseError } from './ParseError.js';
 export { FormatConfig } from './FormatConfig.js';
 export { FormatOptions } from './FormatOptions.js';
+export { stripComments } from './comments.js';